+++ apps/crm/crm/src/app/crm/pages/waterfall/rightholder/rightholder.component.html
@@ -6,6 +6,15 @@
       <mat-label>Waterfall Role</mat-label>
       <mat-error>This field is mandatory</mat-error>
     </static-select>
+    <mat-form-field appearance="outline">
+      <mat-label>Select Locked version</mat-label>
+      <mat-select [formControl]="rightholderForm.get('lockedVersionId')">
+        <mat-option value="">None</mat-option>
+        <mat-option *ngFor="let version of versions" [value]="version.id" >
+          {{ version.name }}
+        </mat-option>
+      </mat-select>
+    </mat-form-field>
     <article>
       <button mat-button color="primary" [disabled]=!rightholderForm.valid (click)="save()">
         <mat-icon svgIcon="edit"></mat-icon>
