+++ apps/crm/crm/src/app/crm/pages/waterfall/rightholder/rightholder.module.ts
@@ -18,6 +18,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatSelectModule } from '@angular/material/select';
 
 @NgModule({
   declarations: [RightholderComponent],
@@ -36,6 +37,7 @@ import { MatSnackBarModule } from '@angular/material/snack-bar';
     MatButtonModule,
     MatFormFieldModule,
     MatSnackBarModule,
+    MatSelectModule,
 
     RouterModule.forChild([{ path: '', component: RightholderComponent }])
   ]
