+++ apps/crm/crm/src/app/crm/pages/waterfall/rightholders/rightholders.component.html
@@ -15,6 +15,9 @@
     <ng-template colRef="roles" label="Roles" let-roles sort>
       {{ roles | toLabel: 'rightholderRoles' }}
     </ng-template>
+    <ng-template colRef="lockedVersionId" label="Locked Version" let-lockedVersionId sort>
+      {{ lockedVersionId ? (lockedVersionId | versionName: waterfall) : '--' }}
+    </ng-template>
     <ng-template colRef="" label="Action" let-rightholder sticky>
       <a mat-icon-button target="_blank" [routerLink]="['/c/o/dashboard/crm/waterfall', waterfall.id, 'rightholders', rightholder.id]" (click)="$event.stopPropagation()">
         <mat-icon svgIcon="launch"></mat-icon>
